// gms-common/node/interactive-analysis/interactive-analysis-ui/packages/weavess/src/ts/components/waveform-display/components/theoretical-phase-window/types.ts
export interface TheoreticalPhaseWindowProps {
  /** Id for window */
  id: string;

  /** Percent location left in div space */
  left: number;

  /** Percent location right in div space */
  right: number;

  /** Lable for window */
  label: string;

  /** color of window */
  color: string;
}

// tslint:disable-next-line:no-empty-interface
export interface TheoreticalPhaseWindowState {
}